chore: toggle 네이밍 변경
