+++ apps/web/src/shared/ui/toggle/index.ts
@@ -1,2 +1,1 @@
-export { Toggle } from './Toggle';
-export type { ToggleProps, ToggleOption } from './Toggle';
+export { Toggle, type ToggleProps } from './Toggle';
