--- apps/web/src/shared/ui/toggleSwitch/index.ts
@@ -1,1 +1,0 @@
-export { ToggleSwitch, type ToggleSwitchProps } from './ToggleSwitch';
