+++ apps/web/src/shared/ui/viewToggle/index.ts
@@ -1,0 +1,2 @@
+export { ViewToggle } from './ViewToggle';
+export type { ViewToggleProps, ViewToggleOption } from './ViewToggle';
